+++ src/_packages/utils.index.ts
@@ -5,6 +5,7 @@ import { extractVariables } from '../conversion/utils/extractVariables';
 import { parseNumber } from '../conversion/utils/parseNumber';
 import { renameParameter } from '../conversion/utils/renameParameter';
 import { titleToName } from '../conversion/utils/titleToName';
+import { forEachAsync } from '../execution/utils/forEachAsync';
 import { replaceParameters } from '../execution/utils/replaceParameters';
 import { CountUtils } from '../utils/expectation-counters';
 import { countCharacters } from '../utils/expectation-counters/countCharacters';
@@ -46,6 +47,8 @@ import { union } from '../utils/sets/union';
 import { trimCodeBlock } from '../utils/trimCodeBlock';
 import { trimEndOfCodeBlock } from '../utils/trimEndOfCodeBlock';
 import { unwrapResult } from '../utils/unwrapResult';
+
+export { forEachAsync };
 
 // TODO: [🌻] For all, decide if theese are internal or external
 export {